test: chessed through the date time problem
